webauthn/createcredential-attachment.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>navigator.credentials.create() authenticator attachment support test</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src=helpers.js></script>
<body></body>
<script>
    // usb transport, with authenticator selection preference
    virtualAuthenticatorPromiseTest(async function() {
        "use strict";
            const credential = await createCredential({
                options: {
                    publicKey: {
                        authenticatorSelection: {
                            authenticatorAttachment: "cross-platform",
                            userVerification: "preferred"
                        },
                    },
                },
            });
            assert_equals(credential.authenticatorAttachment, "cross-platform")
            assert_not_equals(credential.authenticatorAttachment, null);
            assert_not_equals(credential.authenticatorAttachment, "platform");
    }, {
        protocol: "ctap2_1",
        transport: "usb"
    }, "navigator.credentials.create() with usb authenticator, attachment as cross-platform");

    // ble transport, with authenticator selection preference
    virtualAuthenticatorPromiseTest(async function() {
        "use strict";
        const credential = await createCredential({
            options: {
                publicKey: {
                    authenticatorSelection: {
                        authenticatorAttachment: "cross-platform",
                        userVerification: "preferred"
                    },
                },
            },
        });
        assert_equals(credential.authenticatorAttachment, "cross-platform")
        assert_not_equals(credential.authenticatorAttachment, null);
        assert_not_equals(credential.authenticatorAttachment, "platform");
    }, {
        protocol: "ctap2_1",
        transport: "ble"
    }, "navigator.credentials.create() with ble authenticator, attachment as cross-platform");

    // nfc transport, without authenticator selection preference
    virtualAuthenticatorPromiseTest(async function() {
        "use strict";
        const credential = await createCredential({
            options: {
                publicKey: {
                    authenticatorSelection: {
                        userVerification: "preferred"
                    },
                },
            },
        });
        assert_equals(credential.authenticatorAttachment, "cross-platform")
        assert_not_equals(credential.authenticatorAttachment, null);
        assert_not_equals(credential.authenticatorAttachment, "platform");
    }, {
        protocol: "ctap2_1",
        transport: "nfc"
    }, "navigator.credentials.create() with nfc authenticator, attachment as cross-platform");

    // test internal transport, with authenticator selection preference
    virtualAuthenticatorPromiseTest(async function() {
        "use strict";
            const credential = await createCredential({
                options: {
                    publicKey: {
                        authenticatorSelection: {
                            authenticatorAttachment: "platform",
                            userVerification: "preferred"
                        },
                    },
                },
            });
            assert_equals(credential.authenticatorAttachment, "platform")
            assert_not_equals(credential.authenticatorAttachment, null);
            assert_not_equals(credential.authenticatorAttachment, "cross-platform");
    }, {
        protocol: "ctap2_1",
        transport: "internal"
    }, "navigator.credentials.create() with internal authenticator, attachment as platform");

    // test internal transport, without authenticator selection preference
    virtualAuthenticatorPromiseTest(async function() {
        "use strict";
        const credential = await createCredential({
            options: {
                publicKey: {
                    authenticatorSelection: {
                        userVerification: "preferred"
                    },
                },
            },
        });
        assert_equals(credential.authenticatorAttachment, "platform")
        assert_not_equals(credential.authenticatorAttachment, null);
        assert_not_equals(credential.authenticatorAttachment, "cross-platform");
    }, {
        protocol: "ctap2_1",
        transport: "internal"
    }, "navigator.credentials.create() with internal authenticator, attachment as platform inferred");
</script>
